rename component modules name
