+++ src/component/remove.rs
